pyproject.toml
[project]
name = "localllm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "openai>=1.95.1",
    "openapi>=2.0.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
